+++ packages/shared-types/src/api/dtos/users/update-user.dto.ts
@@ -1,6 +1,9 @@
+import { Type } from "class-transformer";
 import {
   IsDate,
   IsEmail,
+  IsObject,
+  IsOptional,
   IsPhoneNumber,
   IsString,
   IsUrl,
@@ -8,15 +11,26 @@ import {
   Matches,
   MaxLength,
   MinLength,
+  ValidateNested,
 } from "class-validator";
 
 import { NAME_REGEX } from "../../../const.constants";
 import { UserIdentifier, UserIdentity } from "../../../user";
 
 export class UpdateUserDto {
+  @IsOptional()
+  @IsObject()
+  @ValidateNested()
+  @Type(() => UpdateIdentifierDto)
   identifier?: UpdateIdentifierDto;
+
+  @IsOptional()
+  @IsObject()
+  @ValidateNested()
+  @Type(() => UpdateIdentityDto)
   identity?: UpdateIdentityDto;
 
+  @IsOptional()
   @IsString()
   @MinLength(6)
   @MaxLength(130)
@@ -27,14 +41,17 @@ class UpdateIdentifierDto
   implements
     Partial<Pick<UserIdentifier, "email" | "phoneNumber" | "username">>
 {
+  @IsOptional()
   @IsString()
   @IsEmail()
   email?: string;
 
+  @IsOptional()
   @IsString()
   @IsPhoneNumber()
   phoneNumber?: string;
 
+  @IsOptional()
   @IsString()
   @MinLength(3)
   username?: string;
@@ -56,6 +73,7 @@ class UpdateIdentityDto
       >
     >
 {
+  @IsOptional()
   @IsString()
   @Length(2, 50)
   @Matches(NAME_REGEX, {
@@ -63,6 +81,7 @@ class UpdateIdentityDto
   })
   firstName?: string;
 
+  @IsOptional()
   @IsString()
   @Length(2, 50)
   @Matches(NAME_REGEX, {
@@ -70,22 +89,28 @@ class UpdateIdentityDto
   })
   lastName?: string;
 
+  @IsOptional()
   @IsString()
   @Length(4, 100)
   displayName?: string;
 
+  @IsOptional()
   @IsString()
   @Length(10, 500)
   description?: string;
 
+  @IsOptional()
   @IsUrl()
   profilePictureUrl?: string | undefined;
 
+  @IsOptional()
   @IsUrl()
   bannerUrl?: string | undefined;
 
+  @IsOptional()
   gender?: string;
 
+  @IsOptional()
   @IsDate()
   birthDate?: Date;
 }